+++ packages/protocolbeat/src/components/editor/lineSelector.ts
