+++ packages/protocolbeat/src/components/editor/useFlagFromQueryParam.ts
@@ -1,3 +1,4 @@
+// No state on purpose - one way only
 export function useQueryParam(key: string): string {
   const url = new URL(window.location.href)
   return url.searchParams.get(key) ?? ''
